requirements.txt
Flask
gunicorn
Keras
numpy
opencv-python
tensorflow-cpu
